+++ packages/wallet-engine/src/engine/approve.ts
@@ -4,20 +4,20 @@ import {
   JsonRpcMiddleware,
   JsonRpcRequest,
   PendingJsonRpcResponse,
-  AsyncJsonRpcEngineNextCallback,
 } from 'json-rpc-engine';
 
 import { Engine } from './engine';
 
 type WithPreopenedInstanceId = {
-  preopenInstanceId: string;
+  preopenInstanceId?: string;
 };
 
-type ProviderRequest<T = unknown> = WithPreopenedInstanceId & {
+type ProviderRequest<T = unknown, U = unknown> = WithPreopenedInstanceId & {
   params: T;
+  proceed: () => Promise<Pick<PendingJsonRpcResponse<U>, 'error' | 'result'>>;
 };
 
-export type PersonalSignRequest = ProviderRequest<[string]>;
+export type PersonalSignRequest = ProviderRequest<[string], string>;
 
 export type IncomingTransaction = {
   from: string;
@@ -30,32 +30,34 @@ export type IncomingTransaction = {
   maxPriorityFeePerGas?: string;
 };
 
-export type SendTransactionRequest = ProviderRequest<[IncomingTransaction]>;
+export type SendTransactionRequest = ProviderRequest<[IncomingTransaction], string>;
 
 export type ApproveEngineOptions = {
   onSendTransaction?: (request: SendTransactionRequest) => Promise<void>;
   onPersonalSign?: (request: PersonalSignRequest) => Promise<void>;
 };
 
-const hasPreopenedPopup = (req: any): req is JsonRpcRequest<unknown> & { preopenInstanceId: string } => {
-  return Object.hasOwn(req, 'preopenInstanceId');
-};
-
 type RequestMiddleware<T, U> = (
-  req: JsonRpcRequest<T> & WithPreopenedInstanceId,
-  res: PendingJsonRpcResponse<U>,
-  next: AsyncJsonRpcEngineNextCallback,
+  req: JsonRpcRequest<T> & Partial<WithPreopenedInstanceId>,
+  proceed: () => Promise<PendingJsonRpcResponse<U>>,
 ) => Promise<void>;
 
 const createRequestMiddleware = <T = any, U = any>(handler: RequestMiddleware<T, U>): JsonRpcMiddleware<any, any> =>
   createAsyncMiddleware<T, U>(async (req, res, next) => {
-    if (!hasPreopenedPopup(req)) {
-      return next();
+    let done = false;
+
+    const proceed = async () => {
+      await next();
+      done = true;
+
+      return res;
+    };
+
+    await handler(req, proceed);
+
+    if (!done) {
+      await proceed();
     }
-
-    await handler(req, res, next);
-
-    next();
   });
 
 const noop = async () => {};
@@ -64,17 +66,19 @@ export const createApproveEngine = ({ onPersonalSign = noop, onSendTransaction =
 
   engine.push(
     createScaffoldMiddleware({
-      personal_sign: createRequestMiddleware<[string]>(async (req, res, next) => {
+      personal_sign: createRequestMiddleware<[string]>(async (req, proceed) => {
         await onPersonalSign({
           preopenInstanceId: req.preopenInstanceId,
           params: req.params!,
+          proceed,
         });
       }),
 
-      eth_sendTransaction: createRequestMiddleware<[IncomingTransaction]>(async (req, res, next) => {
+      eth_sendTransaction: createRequestMiddleware<[IncomingTransaction]>(async (req, proceed) => {
         await onSendTransaction({
           preopenInstanceId: req.preopenInstanceId,
           params: req.params!,
+          proceed,
         });
       }),
     }),