+++ src/components/AccountBalanceWidget/AccountBalanceWidget.tsx
@@ -7,6 +7,7 @@ import { PageHeader } from '../PageHeader';
 import { AddressQRButton } from '../AddressQRButton';
 import { Link } from '../Link';
 import { AddressDropdown } from '../AddressDropdown';
+import { FEATURE_FLAGS } from '~/constants';
 
 export type AccountBalanceWidgetProps = {
   title: string;
@@ -80,17 +81,18 @@ export const AccountBalanceWidget = ({ title, dense = false }: AccountBalanceWid
           </Stack>
 
           <Stack direction="row" justifyContent="flex-end" alignItems="center" spacing={2}>
-            <Button
-              sx={{
-                visibility: 'hidden', // This button is needed just to book the space. Implementation will be added later
-              }}
-              fullWidth={dense}
-              size={dense ? 'medium' : 'large'}
-              variant="contained"
-              startIcon={<TransferIcon />}
-            >
-              Transfer
-            </Button>
+            {FEATURE_FLAGS.transferAssets && (
+              <Button
+                fullWidth={dense}
+                to="transfer"
+                component={Link}
+                size={dense ? 'medium' : 'large'}
+                variant="contained"
+                startIcon={<TransferIcon />}
+              >
+                Transfer
+              </Button>
+            )}
 
             {/* css class "wallet-top-up" is an anchor for product tour */}
             <Button