+++ src/components/Collectibles/CollectibleItemPage.tsx
@@ -1,6 +1,18 @@
-import { Typography, Stack, useIsMobile, Box, NoImageIcon, CopyButton, styled } from '@cere-wallet/ui';
-import { useCollectiblesStore } from '~/hooks/useCollectiblesStore';
+import { useNavigate } from 'react-router-dom';
 import { observer } from 'mobx-react-lite';
+import {
+  Typography,
+  Stack,
+  useIsMobile,
+  Box,
+  NoImageIcon,
+  CopyButton,
+  styled,
+  TransferIcon,
+  Button,
+} from '@cere-wallet/ui';
+
+import { useCollectiblesStore } from '~/hooks';
 import { PageHeader } from '~/components';
 
 type CollectibleItemProps = {
@@ -18,6 +30,7 @@ export const CollectibleItemPage = observer(({ nftId }: CollectibleItemProps) =>
   const collectiblesStore = useCollectiblesStore();
   const nft = collectiblesStore.getNftById(nftId);
   const isMobile = useIsMobile();
+  const navigate = useNavigate();
 
   return (
     <>
@@ -122,10 +135,14 @@ export const CollectibleItemPage = observer(({ nftId }: CollectibleItemProps) =>
               </>
             )}
 
-            {/* TODO functionality Will be added in next iteration */}
-            {/*<Button size="large" variant="contained" startIcon={<TransferIcon />}>*/}
-            {/*  Transfer*/}
-            {/*</Button>*/}
+            <Button
+              size="large"
+              variant="contained"
+              startIcon={<TransferIcon />}
+              onClick={() => navigate({ pathname: '/wallet/home/transfer' })}
+            >
+              Transfer
+            </Button>
           </Stack>
         </Box>
       </Box>